Update getting advanced USB working on new swadge.

+++ components/hdw-usb/advanced_usb_control.c
@@ -7,6 +7,7 @@
 #include <esp_flash.h>
 #include <rom/rtc.h>
 #include <soc/rtc_cntl_reg.h>
+#include <esp_log.h>
 
 #include "tinyusb.h"
 
@@ -188,7 +189,7 @@ void IRAM_ATTR handle_advanced_usb_control_set(int datalen, const uint8_t* data)
                 return;
             }
             intptr_t length = data[6] | (data[7] << 8);
-            ULOG("Writing %d into %p", length, (void*)value);
+            ULOG("Writing %lu into %p", (uint32_t)length, (void*)value);
             memcpy((void*)value, data + 8, length);
             break;
         }
@@ -219,8 +220,8 @@ void IRAM_ATTR handle_advanced_usb_control_set(int datalen, const uint8_t* data)
             // value = -1 will just cause a report.
             // value = 0 will clear it.
             // value < 0 just read current data.
-            ULOG("Allocating to %d (Current: %p / %d)", value, advanced_usb_scratch_buffer_data,
-                 advanced_usb_scratch_buffer_data_size);
+            ULOG("Allocating to %lu (Current: %p / %lu)", (uint32_t)value, advanced_usb_scratch_buffer_data,
+                 (uint32_t)advanced_usb_scratch_buffer_data_size);
             if (!(value & 0x80000000))
             {
                 if (value > advanced_usb_scratch_buffer_data_size)
@@ -244,7 +245,7 @@ void IRAM_ATTR handle_advanced_usb_control_set(int datalen, const uint8_t* data)
             advanced_usb_scratch_immediate[0] = (intptr_t)advanced_usb_scratch_buffer_data;
             advanced_usb_scratch_immediate[1] = advanced_usb_scratch_buffer_data_size;
             advanced_usb_read_offset          = (uint32_t*)(&advanced_usb_scratch_immediate[0]);
-            ULOG("New: %p / %d", advanced_usb_scratch_buffer_data, advanced_usb_scratch_buffer_data_size);
+            ULOG("New: %p / %lu", advanced_usb_scratch_buffer_data, advanced_usb_scratch_buffer_data_size);
             break;
         }
         case ACMD_CMD_MEMSET:
@@ -254,7 +255,7 @@ void IRAM_ATTR handle_advanced_usb_control_set(int datalen, const uint8_t* data)
                 return;
             }
             intptr_t length = data[6] | (data[7] << 8) | (data[8] << 16) | (data[9] << 24);
-            ULOG("Memset %d into %p", length, (void*)value);
+            ULOG("Memset %d into %p", (int)length, (void*)value);
             memset((void*)value, data[10], length);
             break;
         }
